% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-utils.R
\name{assemble_predictions}
\alias{assemble_predictions}
\title{Assemble leave-one-season-out or test phase predictions made by kde, kcde, and sarima
models on training data.}
\usage{
assemble_predictions(preds_path = "inst/estimation/loso-predictions",
  regions = c("National", paste0("Region", 1:10)), seasons = "*",
  models = c("kde", "kcde", "sarima"), prediction_targets = c("onset",
  "peak_week", "peak_inc", "ph_1_inc", "ph_2_inc", "ph_3_inc", "ph_4_inc"),
  prediction_types = c("log_score", "competition_log_score", "bin_log_probs"))
}
\arguments{
\item{preds_path}{path to leave-one-season-out or test phase predictions.}

\item{regions}{character vector specifying regions for which to get predictions,
"National" or "Regionk" for k in 1, ..., 10.}

\item{seasons}{character vector specifying seasons for which to get predictions,
"2011/2012" or "2011-2012" or "*" for all seasons}

\item{models}{character vector specifying models for which to get predictions,
"kde", "kcde", or "sarima"}

\item{prediction_targets}{character vector specifying prediction targets,
"onset", "peak_week", "peak_inc", "ph_1_inc", ..., "ph_4_inc"}

\item{prediction_types}{character vector specifying prediction types,
"log_score", "competition_log_score", or "bin_log_probs"}
}
\value{
a data frame with predictions.
}
\description{
Assemble leave-one-season-out or test phase predictions made by kde, kcde, and sarima
models on training data.
}

