% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-utils.R
\name{assemble_stacking_inputs}
\alias{assemble_stacking_inputs}
\title{Assemble a data frame of inputs to stacking}
\usage{
assemble_stacking_inputs(regions, prediction_target, component_model_names,
  explanatory_variables, include_model_performance = FALSE, preds_path)
}
\arguments{
\item{regions}{string with region: either "National" or in the format
"Regionk" for k in {1, ..., 10}}

\item{prediction_target}{string with either "onset", "peak_week",
"peak_inc", "ph1_inc", ..., "ph4_inc"}

\item{component_model_names}{character vector with names of component models}

\item{explanatory_variables}{character vector with names of explanatory variables
to include for weights; a non-empty subset of "analysis_time_season_week",
"kcde_model_confidence", "sarima_model_confidence", "weighted_ili"}

\item{include_model_performance}{boolean; should measures of model
performance be included in the return result? Generally, should be TRUE if
we're getting inputs for model training and FALSE if we're just getting
covariates to calculate model weights.}

\item{preds_path}{path to directory with leave-one-season-out or test phase
predictions from each component model.  Predictions should be saved in
files named like "kde-National-loso-predictions.rds"}
}
\value{
a data frame with covariates that model weights depend on
  (as specified by explanatory_variables) and possibly log scores from each
  component model
}
\description{
Assemble a data frame of inputs to stacking
}

