% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awes-package.R
\docType{package}
\name{awes}
\alias{awes}
\alias{awes-package}
\title{awes - Adaptively Weighted Ensembles via Stacking}
\description{
Tools for an implementation of adaptively weighted ensembles.
}
\author{
Evan Ray <elray@umass.edu>
}

