% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_median_from_binned_probs}
\alias{calc_median_from_binned_probs}
\title{Calcluation of median value from binned probability distribution}
\usage{
calc_median_from_binned_probs(probs)
}
\arguments{
\item{probs}{vector of named probabilities}
}
\value{
a numeric value
}
\description{
Calcluation of median value from binned probability distribution
}

