% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_competition_log_score}
\alias{compute_competition_log_score}
\title{Calculate "log scores" for the purpose of the competition -- log[sum_i(p_i)] where p_i is the model's
probability of bin i and i runs over some bins adjacent to the bin where the observed quantity was.}
\usage{
compute_competition_log_score(bin_log_probs, observed_bin, prediction_target)
}
\arguments{
\item{bin_log_probs}{named numeric vector with log probability of each bin;
names identify the bins}

\item{observed_bin}{character vector with name(s) of the observed bin(s)
(Note that peak can occur in multiple bins)}

\item{prediction_target}{}
}
\value{
log score for the given observation
}
\description{
Calculate "log scores" for the purpose of the competition -- log[sum_i(p_i)] where p_i is the model's
probability of bin i and i runs over some bins adjacent to the bin where the observed quantity was.
}

