% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{fit_kde_log_peak_week_inc}
\alias{fit_kde_log_peak_week_inc}
\title{Estimate KDE for LOG peak week incidence}
\usage{
fit_kde_log_peak_week_inc(data)
}
\arguments{
\item{data}{}
}
\value{
a fit from density(), along with points to evaluate at
}
\description{
Estimate KDE for LOG peak week incidence
}

