% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{fit_kde_peak_week}
\alias{fit_kde_peak_week}
\title{Estimate KDE for peak week}
\usage{
fit_kde_peak_week(data)
}
\arguments{
\item{data}{}
}
\value{
a fit from density(), along with points to evaluate at
}
\description{
Estimate KDE for peak week
}

