% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{fit_kde_weekly_inc}
\alias{fit_kde_weekly_inc}
\title{Fit a GAM for weekly incidence}
\usage{
fit_kde_weekly_inc(data)
}
\arguments{
\item{data}{}
}
\value{
a mgcv smooth spline fit
}
\description{
Fit a GAM for weekly incidence
}

