% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-utils.R
\name{fit_kernel_smoothed_stacked_model_fixed_bw}
\alias{fit_kernel_smoothed_stacked_model_fixed_bw}
\title{Estimate weights for a linear combination of predictive models.}
\usage{
fit_kernel_smoothed_stacked_model_fixed_bw(component_model_log_scores,
  covariate, prediction_covariate, bw, tol = .Machine$double.eps)
}
\arguments{
\item{component_model_log_scores}{a data frame or matrix of log scores.
Each column gives log scores for a particular predictive model.
Each row corresponds to one observation}

\item{covariate}{a single covariate over which weights should be smoothed
(more than one covariate may be supported later)}

\item{bw}{a bandwidth for the smoothing}

\item{tol}{numeric, if method was "em", stopping tolerance}
}
\value{
model weights
}
\description{
This function uses the "degenerate EM" algorithm outlined at
http://www.cs.cmu.edu/~roni/11761-f16/Presentations/degenerateEM.pdf,
but modified in a way that is ad hoc but I think could be justified
to use kernel-weighted observations.
}

