% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{fit_region_kdes}
\alias{fit_region_kdes}
\title{Fits and saves KDE models for all region-year combinations}
\usage{
fit_region_kdes(data, region, first_test_year, first_test_week, path)
}
\arguments{
\item{data}{cleaned usflu dataset}

\item{region}{character string identifying a region}

\item{first_test_year}{along with first_test_week, defines the first time at which data is left out}

\item{first_test_week}{}

\item{path}{filepath for saving}
}
\value{
nothing, just saving files
}
\description{
Fits and saves KDE models for all region-year combinations
}

