% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima-utils.R
\name{fit_region_sarima}
\alias{fit_region_sarima}
\title{Estimate series of leave-one-season-out SARIMA models}
\usage{
fit_region_sarima(data, reg_num, first_test_season, d = NA, D = NA,
  seasonal_difference = TRUE, transformation = "none", path)
}
\arguments{
\item{data}{regional dataset with structure like regionflu-cleaned}

\item{reg_num}{region number for estimation}

\item{first_test_season}{string indicating first test season}

\item{d}{order of first differencing}

\item{D}{order of seasonal differencing}

\item{seasonal_difference}{boolean; take a seasonal difference before passing
to auto.arima?}

\item{transformation}{character specifying transformation type:
"box-cox", "log", or "none"}

\item{path}{path in which to save files}
}
\value{
NULL just saves files
}
\description{
Estimate series of leave-one-season-out SARIMA models
}

