% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu_data}
\alias{flu_data}
\title{Influenza incidence in the US (1997 - 2016)}
\format{A data.frame with 11,055 observations on weighted influenza-like illness 
measurements from all HHS regions, including the national level.}
\source{
The cdcfluview R package.
}
\usage{
flu_data
}
\description{
A dataset of public influenza data from the US CDC.
}
\keyword{datasets}

