% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu_onset_baselines}
\alias{flu_onset_baselines}
\title{Flu season "onset thresholds" from the US CDC.}
\format{A data.frame with 220 observations of the seasonal baseline threshold used for 
determining the "flu season onset" in each region of the US.}
\source{
\url{https://github.com/cdcepi/FluSight-forecasts/blob/master/wILI_Baseline.csv}
}
\usage{
flu_onset_baselines
}
\description{
Flu season "onset thresholds" from the US CDC.
}
\keyword{datasets}

