% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_initial_rng_substream}
\alias{get_initial_rng_substream}
\title{Get the initial rng substream for an rstream object.  Should be called by all
prediction methods before doing any random number generation.}
\usage{
get_initial_rng_substream(seed = 4114043)
}
\arguments{
\item{seed}{integer seed for rng; the default was randomly generated}
}
\value{
object of class "rstream.mrg32k3a".  The object has been packed via
  rstream.packed.
}
\description{
This function DOES NOT set RNG to use rstream with the returned object.
Because of strange behavior in the rstream package, this function can be
called only once in a given R session.
}

