% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{get_log_scores_via_direct_simulation}
\alias{get_log_scores_via_direct_simulation}
\title{Get log scores and full predictive distributions for each prediction target}
\usage{
get_log_scores_via_direct_simulation(analysis_time_season,
  first_analysis_time_season_week = 10, last_analysis_time_season_week = 41,
  region, prediction_target_var, incidence_bins, incidence_bin_names, n_sims,
  model_name, fits_path, prediction_save_path)
}
\arguments{
\item{analysis_time_season}{character vector of length 1 specifying the
season to obtain predictions for, in the format "2000/2001"}

\item{first_analysis_time_season_week}{integer specifying the first week of
the season in which to make predictions, using all data up to and
including that week to make predictions for each following week in the
season}

\item{last_analysis_time_season_week}{integer specifying the last week of
the season in which to make predictions, using all data up to and including
that week to make predictions for each following week in the season}

\item{region}{string specifying the region to use, in the format "X" or
"Region k" where k \in {1, ..., 10}}

\item{prediction_target_var}{string specifying the name of the variable in
data for which we want to make predictions}

\item{incidence_bins}{a data frame with variables lower and upper defining
lower and upper endpoints to use in binning incidence}

\item{incidence_bin_names}{a character vector with a name for each incidence
bin}

\item{n_sims}{integer number of samples to simulate}

\item{model_name}{name of model, stored in the results data frame}

\item{fits_path}{path to directory where fitted models are stored}

\item{prediction_save_path}{path to directory where results will be saved}
}
\value{
none
}
\description{
for a given season using a predictive method that works by directly 
simulating predictive distributions of each target.  Results are
stored in a data frame, saved in a .rds file with a name like
"model_name-region-season-loso-predictions.rds"
Results have columns indicating the analysis time season and season week,
model name, log scores for each prediction target, the "log score" used
in the competition (adding probabilities from adjacent bins) for each
prediction target, as well as the log of the probability assigned to each
bin.
}

