% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_observed_seasonal_quantities}
\alias{get_observed_seasonal_quantities}
\title{Compute season onset, peak week, and peak incidence}
\usage{
get_observed_seasonal_quantities(data, season, first_CDC_season_week = 10,
  last_CDC_season_week = 42, onset_baseline, incidence_var, incidence_bins,
  incidence_bin_names)
}
\arguments{
\item{data}{a data frame containing at minimum columns named season, 
season_week and a column with some sort of incidence measure}

\item{season}{the season to look at}

\item{first_CDC_season_week}{the first week of the season to use for
calculating onset and peak}

\item{last_CDC_season_week}{the last week of the season to use for
calculating onset and peak}

\item{onset_baseline}{numeric baseline value for determining season onset}

\item{incidence_var}{a character string naming the variable in the data
argument containing a measure of incidence, or an integer index}

\item{incidence_bins}{a data frame with variables lower and upper defining
lower and upper endpoints to use in binning incidence}

\item{incidence_bin_names}{a character vector with a name for each incidence
bin}
}
\value{
a list with four entries:
  1) observed_onset_week, either an integer between first_CDC_season_week
    and last_CDC_season_week (inclusive), or "none"
  2) observed_peak_week, an integer between first_CDC_season_week and
    last_CDC_season_week (inclusive)
  3) observed_peak_inc, a numeric with the maximum value of the specified
    incidence measure between first_CDC_season_week and last_CDC_season_week
  4) observed_peak_inc_bin, character name of incidence bin for peak incidence
}
\description{
Compute season onset, peak week, and peak incidence
}

