% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_onset_baseline}
\alias{get_onset_baseline}
\title{Get the onset baseline for a combination of region and season}
\usage{
get_onset_baseline(region, season = "2015/2016")
}
\arguments{
\item{region}{a string, either "X", "Region k", or "Regionk" where
k \in {1, ..., 10}}

\item{season}{a string, in the format "2015/2016"}
}
\value{
baseline value for determining season onset
}
\description{
Get the onset baseline for a combination of region and season
}

