% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rng_substream}
\alias{get_rng_substream}
\title{Get the rng substream for an rstream object corresponding to the combination
of prediction method, region, and season left out.  Should be called by all
prediction methods before doing any random number generation.}
\usage{
get_rng_substream(rngstream, seed = 4114043, method, region, season,
  set_rng = TRUE)
}
\arguments{
\item{rngstream}{(optional) object of class "rstream.mrg32k3a" which will be
advanced from its current state.}

\item{seed}{integer seed for rng; the default was randomly generated}

\item{method}{character string specifying prediction method
currently one of "sarima", "kcde", or "kde"}

\item{region}{character string specifying region; format "National" or
"Region k" for k \in {1, ..., 10}}

\item{season}{character string specifying season predicted, in format
"1998/1999"}

\item{set_rng}{boolean should rng be set to use rstream with the returned
rngstream object?}
}
\value{
invisible object of class "rstream.mrg32k3a", but advanced to the
first substream reserved for the given combination of prediction method,
region, and season.  The object has been packed via rstream.packed.
}
\description{
Importantly, by default this function has the side effect of setting RNG to
use rstream with the returned object.  This behavior is determined by the
set_rng argument.  This means the caller doesn't have to worry about doing
anything unless (a) it wants to use more than 1 substream or (b) it is going
to parallelize or do any RNG in a different R session.  Because of strange
behavior in the rstream package, this function can be called at most once
without the rngstream argument in a given R session.
}

