% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logspace_add}
\alias{logspace_add}
\title{Calculate log(exp(logx) + exp(logy)) in a somewhat numerically stable way.}
\usage{
logspace_add(logx, logy)
}
\arguments{
\item{logx, }{logy log-scale numeric values to add}
}
\value{
log(exp(logx) + exp(logy)), but more numerically stable
}
\description{
Calculate log(exp(logx) + exp(logy)) in a somewhat numerically stable way.
}

