% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logspace_sub_matrix_rows}
\alias{logspace_sub_matrix_rows}
\title{Calculate logspace difference of matrix rows in a somewhat numerically stable
way.}
\usage{
logspace_sub_matrix_rows(logX)
}
\arguments{
\item{logX}{log-scale numeric matrix of values to subtract.  logX must have
exactly 2 columns.}
}
\value{
log(exp(logX)[, 1] - exp(logX)[, 2]), but more numerically stable
}
\description{
Calculate logspace difference of matrix rows in a somewhat numerically stable
way.
}

