% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logspace_sum_matrix_rows}
\alias{logspace_sum_matrix_rows}
\title{Calculate logspace summation of matrix rows in a somewhat numerically stable
way.}
\usage{
logspace_sum_matrix_rows(logX)
}
\arguments{
\item{logX}{log-scale numeric matrix of values to sum.}
}
\value{
log(apply(exp(logX), 1, sum)), but more numerically stable
}
\description{
Calculate logspace summation of matrix rows in a somewhat numerically stable
way.
}

