% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_predictions_dataframe}
\alias{make_predictions_dataframe}
\title{Helper function to standardize creation of empty prediction dataset}
\usage{
make_predictions_dataframe(data, model_name, incidence_bin_names,
  first_analysis_time_season_week = 10, last_analysis_time_season_week = 41)
}
\arguments{
\item{data}{the data to be used}

\item{model_name}{the name of the model}

\item{incidence_bin_names}{character vector of names for incidence bins}

\item{first_analysis_time_season_week}{the first week during each season in
which we will form a prediction}

\item{last_analysis_time_season_week}{the last week during each season in
which we will form a prediction}
}
\value{
the prediction_df filled with NAs
}
\description{
Helper function to standardize creation of empty prediction dataset
}

