% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-prediction-utils.R
\name{predict_kde_log_peak_week_inc}
\alias{predict_kde_log_peak_week_inc}
\title{Compute predictive distribution for peak week incidence from KDE fit}
\usage{
predict_kde_log_peak_week_inc(fm, bins, bin_names, n_sim)
}
\arguments{
\item{fm}{output from fit_kde_log_peak_week_inc()}

\item{bins}{cutpoints for incidence bins}

\item{bin_names}{vector of bin names, length 1 fewer than length(bins)}

\item{n_sim}{number of draws from predictive distribution to return}
}
\value{
vector of probabilities representing the predictive distribution of NOT LOG peak week incidence, 
        in given bins
}
\description{
Compute predictive distribution for peak week incidence from KDE fit
}

