% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-prediction-utils.R
\name{predict_kde_log_weekly_inc}
\alias{predict_kde_log_weekly_inc}
\title{Compute predictive distribution for weekly incidence from GAM fit}
\usage{
predict_kde_log_weekly_inc(fm, season_weeks, bins, bin_names, n_sim)
}
\arguments{
\item{fm}{a GAM fit for weekly incidence}

\item{season_weeks}{season_weeks to predict for}

\item{bins}{cutpoints for incidence bins}

\item{bin_names}{vector of bin names, length 1 fewer than length(bins)}

\item{n_sim}{number of draws from predictive distribution to return}
}
\value{
matrix of probabilities representing the predictive distribution of NOT LOG scale weekly incidence, 
        with row correponding to the bin and column corresponding to each given season_week
}
\description{
Compute predictive distribution for weekly incidence from GAM fit
}

