% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-prediction-utils.R
\name{predict_kde_peak_week}
\alias{predict_kde_peak_week}
\title{Compute predictive distribution for peak week from KDE fit}
\usage{
predict_kde_peak_week(fm, n_sim)
}
\arguments{
\item{fm}{output from fit_kde_peak_week()}

\item{n_sim}{number of draws from predictive distribution to return}
}
\value{
vector of 52 probabilities representing the predictive distribution of peak week, 
        in bins representing season week 1 through 52 (or EW31 through EW30 -- or EW29 if a 53 week year) 
        with the final entry representing the probability of no onset#'
}
\description{
Compute predictive distribution for peak week from KDE fit
}

