% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{predict_region_kde}
\alias{predict_region_kde}
\title{Wrapper for prediction and evaluation of KDE fits}
\usage{
predict_region_kde(data, region, path, n_sim)
}
\arguments{
\item{data}{data, to be used for evaluation}

\item{region}{region of focus}

\item{path}{filepath to fitted models}

\item{n_sim}{number of simulations to run for predictive distributions}
}
\value{
NULL just saves a file
}
\description{
The function follows this outline
\itemize{
      \item{determine set of years for which LOSO fits are available}
      \item{For those years, generate predictions made at each season-week. 
            These will be the same predictions repeated, as models not dependent on any inputs.}
      \item{Compare predictions to the real data}
      \item{Save object with comparisons}
}
}

