% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima-utils.R
\name{predict_region_sarima}
\alias{predict_region_sarima}
\title{Create regional predictions from a seasonal model}
\usage{
predict_region_sarima(data, reg_num, first_test_season, n_sim, filepath)
}
\arguments{
\item{data}{regionalflu-cleaned dataset}

\item{reg_num}{region number}

\item{first_test_season}{the first season reserved for testing data. seasons before this are predicted.}

\item{n_sim}{number of draws to simulate from each predictive distribution.}

\item{filepath}{path to folder with estimated fits}
}
\value{
NULL just saves a file
}
\description{
Create regional predictions from a seasonal model
}

