% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima-utils.R
\name{sample_predictive_trajectories_arima}
\alias{sample_predictive_trajectories_arima}
\title{Simulate predictive trajectories from an ARIMA model}
\usage{
sample_predictive_trajectories_arima(object, h = ifelse(object$arma[5] > 1, 2
  * object$arma[5], 10), level = c(80, 95), fan = FALSE, xreg = NULL,
  lambda = object$lambda, npaths = 5000, ...)
}
\arguments{
\item{object}{an Arima fit object (with class "Arima")}

\item{h}{number of time steps forwards to simulate}

\item{level}{not used}

\item{fan}{not used}

\item{xreg}{not used}

\item{lambda}{not used}

\item{npaths}{number of sample trajectories to simulate}
}
\value{
an npaths by h matrix with simulated values
}
\description{
This is directly taken from the forecast.Arima function from the forecast package,
but I've forced bootstrap = TRUE and return the simulated trajectories which were
not returned in the original function definition.
}

