% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{season_week_to_year_week}
\alias{season_week_to_year_week}
\title{Convert season week to year week}
\usage{
season_week_to_year_week(season_week, first_season_week = 31,
  weeks_in_first_season_year)
}
\arguments{
\item{season_week}{vector of indices of weeks in season}

\item{first_season_week}{number of week in year corresponding to the first
week in the season.  For example, our code takes this value to be 31:
a new influenza season starts on the 31st week of each year.}

\item{weeks_in_first_season_year}{How many MMWR weeks are in the first year
of the season?  For example, in the 2000/2001 season, the first year is
2000.  There were 52 MMWR weeks in 2000.}
}
\value{
vector of the same length as season_week with the week of the year
  that each observation falls on
}
\description{
Convert season week to year week
}

