% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcde-utils.R
\name{simulate_trajectories_kcde}
\alias{simulate_trajectories_kcde}
\title{Simulate predictive distributions from KCDE/copulas model; suitable for use
as the \code{simulate_trajectories_function} argument to
\code{get_log_scores_via_trajectory_simulation}.}
\usage{
simulate_trajectories_kcde(n_sims, max_prediction_horizon, data, region,
  analysis_time_season, analysis_time_season_week, params)
}
\arguments{
\item{n_sims}{number of trajectories to simulate}

\item{max_prediction_horizon}{how many steps ahead to simulate}

\item{data}{data set}

\item{region}{region}

\item{analysis_time_season}{season in which we're predicting}

\item{analysis_time_season_week}{week of the season in which we're making our
predictions, using all data up to the analysis time to make predictions for
later time points}

\item{params}{other parameters.  A list with the following entries:
* n_kcde_sims = number of simulations from kcde predictive distributions;
    this is different from number of predictive trajectories
* copula_save_path = path to directory where copula fits are stored
* estimation_results_path = path to directory where kcde fits are stored
* max_lag = character giving maximum lag value used in kcde fits
* seasonality = logical giving whether a seasonal kernel was used in kcde
* bw_parameterization = string giving parameterization of bandwidth in kcde
    fit
* last_analysis_time_season_week = 41}
}
\value{
an n_sims by h matrix with simulated values
}
\description{
Simulate predictive distributions from KCDE/copulas model; suitable for use
as the \code{simulate_trajectories_function} argument to
\code{get_log_scores_via_trajectory_simulation}.
}

