% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-utils.R
\name{weighted_combine_predictions}
\alias{weighted_combine_predictions}
\title{Compute a weighted combination of prediction results}
\usage{
weighted_combine_predictions(component_preds, weights, log_weights)
}
\arguments{
\item{component_preds}{data frame of predictions from component models}

\item{weights}{data frame of model weights}

\item{log_weights}{data frame of (log) model weights}
}
\value{
data frame with weighted combination of predictions
}
\description{
Compute a weighted combination of prediction results
}

