% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{year_week_to_season_week}
\alias{year_week_to_season_week}
\title{Convert year week to season week}
\usage{
year_week_to_season_week(year_week, year)
}
\arguments{
\item{year_week}{vector of indices of weeks in year}

\item{year}{either a single (four digit) year or a vector of years with the
same length as year_week}
}
\value{
vector of the same length as year_week with the week of the season
  that each observation falls on
}
\description{
Convert year week to season week
}

